+++ terraform/python/rekognition_api/conf.py
@@ -15,10 +15,12 @@ to display the configuration values in the /info endpoint.
 
 """
 
+
 import importlib.util
-import logging
 
 # python stuff
+import json
+import logging
 import os  # library for interacting with the operating system
 import platform  # library to view information about the server host this Lambda runs on
 import re
@@ -26,6 +28,8 @@ from typing import Any, Dict, List, Optional
 
 # 3rd party stuff
 import boto3  # AWS SDK for Python https://boto3.amazonaws.com/v1/documentation/api/latest/index.html
+import pkg_resources
+from botocore.exceptions import NoCredentialsError, ProfileNotFound
 from dotenv import load_dotenv
 from pydantic import Field, SecretStr, ValidationError, ValidationInfo, field_validator
 from pydantic_settings import BaseSettings
@@ -90,6 +94,9 @@ class SettingsDefaults:
     DEBUG_MODE: bool = bool(TFVARS.get("debug_mode", False))
     SHARED_RESOURCE_IDENTIFIER = TFVARS.get("shared_resource_identifier", "rekognition_api")
 
+    AWS_APIGATEWAY_CREATE_CUSTOM_DOMAIN = TFVARS.get("aws_apigateway_create_custom_domaim", False)
+    AWS_APIGATEWAY_ROOT_DOMAIN = TFVARS.get("aws_apigateway_root_domain", None)
+
     AWS_DYNAMODB_TABLE_ID = "rekognition"
 
     AWS_REKOGNITION_COLLECTION_ID = AWS_DYNAMODB_TABLE_ID + "-collection"
@@ -131,15 +138,21 @@ def empty_str_to_int_default(v: str, default: int) -> int:
 
 
 # pylint: disable=too-many-public-methods
+# pylint: disable=too-many-instance-attributes
 class Settings(BaseSettings):
     """Settings for Lambda functions"""
 
     _aws_session: boto3.Session = None
+    _aws_apigateway_client = None
+    _aws_s3_client = None
+    _aws_dynamodb_client = None
+    _aws_rekognition_client = None
     _aws_access_key_id_source: str = "unset"
     _aws_secret_access_key_source: str = "unset"
     _dump: dict = None
     _initialized: bool = False
 
+    # pylint: disable=too-many-branches
     def __init__(self, **data: Any):
         super().__init__(**data)
         if bool(os.environ.get("AWS_DEPLOYED", False)):
@@ -148,30 +161,55 @@ class Settings(BaseSettings):
             self._aws_secret_access_key_source: str = "overridden by IAM role-based security"
             self._aws_session = boto3.Session()
             self._initialized = True
-            return
 
-        aws_profile = str(os.environ.get("AWS_PROFILE", "")).strip()
-        if len(aws_profile) > 0:
-            logger.debug("Using AWS_PROFILE: %s", aws_profile)
-            self._aws_access_key_id_source = "aws_profile"
-            self._aws_secret_access_key_source = "aws_profile"
+        if not self._initialized and bool(os.environ.get("GITHUB_ACTIONS", False)):
+            try:
+                self._aws_session = boto3.Session(
+                    region_name=os.environ.get("AWS_REGION", "us-east-1"),
+                    aws_access_key_id=os.environ.get("AWS_ACCESS_KEY_ID", None),
+                    aws_secret_access_key=os.environ.get("AWS_SECRET_ACCESS_KEY", None),
+                )
+            except ProfileNotFound:
+                logger.warning("aws_profile %s not found", self.aws_profile)
+
+            if self.aws_profile:
+                self._aws_access_key_id_source = "aws_profile"
+                self._aws_secret_access_key_source = "aws_profile"
+            else:
+                self._aws_access_key_id_source = "environ"
+                self._aws_secret_access_key_source = "environ"
             self._initialized = True
-            return
 
-        if "aws_access_key_id" in data or "aws_secret_access_key" in data:
-            if "aws_access_key_id" in data:
-                self._aws_access_key_id_source = "constructor"
-            if "aws_secret_access_key" in data:
-                self._aws_secret_access_key_source = "constructor"
-            self._initialized = True
-            return
+        if not self._initialized:
+            if self.aws_profile:
+                self._aws_access_key_id_source = "aws_profile"
+                self._aws_secret_access_key_source = "aws_profile"
+                self._initialized = True
 
-        if "AWS_ACCESS_KEY_ID" in os.environ:
-            self._aws_access_key_id_source = "environ"
-        if "AWS_SECRET_ACCESS_KEY" in os.environ:
-            self._aws_secret_access_key_source = "environ"
+        if not self._initialized:
+            if "aws_access_key_id" in data or "aws_secret_access_key" in data:
+                if "aws_access_key_id" in data:
+                    self._aws_access_key_id_source = "constructor"
+                if "aws_secret_access_key" in data:
+                    self._aws_secret_access_key_source = "constructor"
+                self._initialized = True
+
+        if not self._initialized:
+            if "AWS_ACCESS_KEY_ID" in os.environ:
+                self._aws_access_key_id_source = "environ"
+            if "AWS_SECRET_ACCESS_KEY" in os.environ:
+                self._aws_secret_access_key_source = "environ"
+
+        if self.debug_mode:
+            logger.setLevel(logging.DEBUG)
+
+        # pylint: disable=logging-fstring-interpolation
+        logger.debug(f"initialized settings: {self.aws_auth}")
         self._initialized = True
 
+    shared_resource_identifier: Optional[str] = Field(
+        SettingsDefaults.SHARED_RESOURCE_IDENTIFIER, env="SHARED_RESOURCE_IDENTIFIER"
+    )
     debug_mode: Optional[bool] = Field(
         SettingsDefaults.DEBUG_MODE,
         env="DEBUG_MODE",
@@ -200,6 +238,16 @@ class Settings(BaseSettings):
     aws_region: Optional[str] = Field(
         SettingsDefaults.AWS_REGION,
         env="AWS_REGION",
+    )
+    aws_apigateway_create_custom_domaim: Optional[bool] = Field(
+        SettingsDefaults.AWS_APIGATEWAY_CREATE_CUSTOM_DOMAIN,
+        env="AWS_APIGATEWAY_CREATE_CUSTOM_DOMAIN",
+        pre=True,
+        getter=lambda v: empty_str_to_bool_default(v, SettingsDefaults.AWS_APIGATEWAY_CREATE_CUSTOM_DOMAIN),
+    )
+    aws_apigateway_root_domain: Optional[str] = Field(
+        SettingsDefaults.AWS_APIGATEWAY_ROOT_DOMAIN,
+        env="AWS_APIGATEWAY_ROOT_DOMAIN",
     )
     aws_dynamodb_table_id: Optional[str] = Field(
         SettingsDefaults.AWS_DYNAMODB_TABLE_ID,
@@ -232,9 +280,11 @@ class Settings(BaseSettings):
         pre=True,
         getter=lambda v: empty_str_to_int_default(v, SettingsDefaults.AWS_REKOGNITION_FACE_DETECT_THRESHOLD),
     )
-    shared_resource_identifier: Optional[str] = Field(
-        SettingsDefaults.SHARED_RESOURCE_IDENTIFIER, env="SHARED_RESOURCE_IDENTIFIER"
-    )
+
+    @property
+    def aws_account_id(self):
+        """AWS account id"""
+        return self.aws_session.client("sts").get_caller_identity()["Account"]
 
     @property
     def aws_access_key_id_source(self):
@@ -247,12 +297,26 @@ class Settings(BaseSettings):
         return self._aws_secret_access_key_source
 
     @property
+    def aws_auth(self) -> dict:
+        """AWS authentication"""
+        return {
+            "aws_profile": self.aws_profile,
+            "aws_access_key_id_source": self.aws_access_key_id_source,
+            "aws_secret_access_key_source": self.aws_secret_access_key_source,
+            "aws_region": self.aws_region,
+        }
+
+    @property
     def aws_session(self):
         """AWS session"""
         if not self._aws_session:
             if self.aws_profile:
                 logger.debug("creating new aws_session with aws_profile: %s", self.aws_profile)
-                self._aws_session = boto3.Session(profile_name=self.aws_profile, region_name=self.aws_region)
+                try:
+                    self._aws_session = boto3.Session(profile_name=self.aws_profile, region_name=self.aws_region)
+                except ProfileNotFound:
+                    logger.warning("aws_profile %s not found", self.aws_profile)
+
                 return self._aws_session
             if self.aws_access_key_id.get_secret_value() is not None and self.aws_secret_access_key is not None:
                 logger.debug("creating new aws_session with aws keypair: %s", self.aws_access_key_id_source)
@@ -267,24 +331,60 @@ class Settings(BaseSettings):
         return self._aws_session
 
     @property
-    def s3_client(self):
+    def aws_apigateway_client(self):
+        """API Gateway client"""
+        if not self._aws_apigateway_client:
+            self._aws_apigateway_client = self.aws_session.client("apigateway")
+        return self._aws_apigateway_client
+
+    @property
+    def aws_s3_client(self):
         """S3 client"""
-        return self.aws_session.resource("s3")
+        if not self._aws_s3_client:
+            self._aws_s3_client = self.aws_session.client("s3")
+        return self._aws_s3_client
 
     @property
-    def dynamodb_client(self):
+    def aws_dynamodb_client(self):
         """DynamoDB client"""
-        return self.aws_session.resource("dynamodb")
+        if not self._aws_dynamodb_client:
+            self._aws_dynamodb_client = self.aws_session.client("dynamodb")
+        return self._aws_dynamodb_client
 
     @property
-    def rekognition_client(self):
+    def aws_rekognition_client(self):
         """Rekognition client"""
-        return self.aws_session.client("rekognition")
+        if not self._aws_rekognition_client:
+            self._aws_rekognition_client = self.aws_session.client("rekognition")
+        return self._aws_rekognition_client
 
     @property
     def dynamodb_table(self):
         """DynamoDB table"""
-        return self.dynamodb_client.Table(self.aws_dynamodb_table_id)
+        return self.aws_dynamodb_client.Table(self.aws_dynamodb_table_id)
+
+    @property
+    def aws_s3_bucket_name(self) -> str:
+        """Return the S3 bucket name."""
+        return self.aws_account_id + "-" + self.shared_resource_identifier
+
+    @property
+    def aws_apigateway_name(self) -> str:
+        """Return the API name."""
+        return self.shared_resource_identifier + "-api"
+
+    @property
+    def aws_apigateway_domain_name(self) -> str:
+        """Return the API domain."""
+        if self.aws_apigateway_create_custom_domaim:
+            return "api." + self.shared_resource_identifier + "." + self.aws_apigateway_root_domain
+
+        response = self.aws_apigateway_client.get_rest_apis()
+        for item in response["items"]:
+            if item["name"] == self.aws_apigateway_name:
+                api_id = item["id"]
+                return f"{api_id}.execute-api.{settings.aws_region}.amazonaws.com"
+        return None
 
     @property
     def is_using_dotenv_file(self) -> bool:
@@ -328,8 +428,21 @@ class Settings(BaseSettings):
         def recursive_sort_dict(d):
             return {k: recursive_sort_dict(v) if isinstance(v, dict) else v for k, v in sorted(d.items())}
 
+        def get_installed_packages():
+            installed_packages = pkg_resources.working_set
+            # pylint: disable=not-an-iterable
+            package_list = [(d.project_name, d.version) for d in installed_packages]
+            return package_list
+
         if self._dump and self._initialized:
             return self._dump
+
+        if not self._initialized:
+            return {}
+
+        packages = get_installed_packages()
+        packages_dict = [{"name": name, "version": version} for name, version in packages]
+        packages_json = json.dumps(packages_dict, indent=4)
 
         self._dump = {
             "environment": {
@@ -345,22 +458,38 @@ class Settings(BaseSettings):
                 "debug_mode": self.debug_mode,
                 "dump_defaults": self.dump_defaults,
                 "version": self.version,
+                "python_version": platform.python_version(),
+                "python_implementation": platform.python_implementation(),
+                "python_compiler": platform.python_compiler(),
+                "python_build": platform.python_build(),
+                "python_branch": platform.python_branch(),
+                "python_revision": platform.python_revision(),
+                "python_version_tuple": platform.python_version_tuple(),
+                "python_installed_packages": packages_json,
             },
-            "aws": {
-                "aws_profile": self.aws_profile,
-                "aws_access_key_id_source": self.aws_access_key_id_source,
-                "aws_secret_access_key_source": self.aws_secret_access_key_source,
-                "aws_region": self.aws_region,
-            },
-            "rekognition": {
+            "aws_auth": self.aws_auth,
+            "aws_rekognition": {
                 "aws_rekognition_collection_id": self.aws_rekognition_collection_id,
                 "aws_rekognition_face_detect_max_faces_count": self.aws_rekognition_face_detect_max_faces_count,
                 "aws_rekognition_face_detect_attributes": self.aws_rekognition_face_detect_attributes,
                 "aws_rekognition_face_detect_quality_filter": self.aws_rekognition_face_detect_quality_filter,
                 "aws_rekognition_face_detect_threshold": self.aws_rekognition_face_detect_threshold,
             },
-            "dynamodb": {
+            "aws_dynamodb": {
                 "aws_dynamodb_table_id": self.aws_dynamodb_table_id,
+            },
+            "aws_apigateway": {
+                "aws_apigateway_create_custom_domaim": self.aws_apigateway_create_custom_domaim,
+                "aws_apigateway_name": self.aws_apigateway_name,
+                "aws_apigateway_root_domain": self.aws_apigateway_root_domain,
+                "aws_apigateway_domain_name": self.aws_apigateway_domain_name,
+            },
+            "aws_lambda": {},
+            "aws_s3": {
+                "aws_s3_bucket_prefix": self.aws_s3_bucket_name,
+            },
+            "terraform": {
+                "tfvars": self.tfvars_variables,
             },
         }
         if self.dump_defaults:
@@ -402,8 +531,10 @@ class Settings(BaseSettings):
             v = SecretStr(v)
         if v.get_secret_value() in [None, ""]:
             return SettingsDefaults.AWS_ACCESS_KEY_ID
-        if "aws_profile" in values.data and values.data["aws_profile"] != SettingsDefaults.AWS_PROFILE:
-            logger.warning("aws_access_key_id is ignored when aws_profile is set")
+        aws_profile = values.data.get("aws_profile", None)
+        if aws_profile and len(aws_profile) > 0 and aws_profile != SettingsDefaults.AWS_PROFILE:
+            # pylint: disable=logging-fstring-interpolation
+            logger.warning(f"aws_access_key_id is being ignored. using aws_profile {aws_profile}.")
             return SettingsDefaults.AWS_ACCESS_KEY_ID
         return v
 
@@ -414,8 +545,10 @@ class Settings(BaseSettings):
             v = SecretStr(v)
         if v.get_secret_value() in [None, ""]:
             return SettingsDefaults.AWS_SECRET_ACCESS_KEY
-        if "aws_profile" in values.data and values.data["aws_profile"] != SettingsDefaults.AWS_PROFILE:
-            logger.warning("aws_secret_access_key is ignored when aws_profile is set")
+        aws_profile = values.data.get("aws_profile", None)
+        if aws_profile and len(aws_profile) > 0 and aws_profile != SettingsDefaults.AWS_PROFILE:
+            # pylint: disable=logging-fstring-interpolation
+            logger.warning(f"aws_secret_access_key is being ignored. using aws_profile {aws_profile}.")
             return SettingsDefaults.AWS_SECRET_ACCESS_KEY
         return v
 
@@ -428,6 +561,20 @@ class Settings(BaseSettings):
             return SettingsDefaults.AWS_REGION
         if v not in valid_regions:
             raise RekognitionValueError(f"aws_region {v} not in aws_regions")
+        return v
+
+    @field_validator("aws_apigateway_root_domain")
+    def validate_aws_apigateway_root_domain(cls, v) -> str:
+        """Validate aws_apigateway_root_domain"""
+        if v in [None, ""]:
+            return SettingsDefaults.AWS_APIGATEWAY_ROOT_DOMAIN
+        return v
+
+    @field_validator("aws_apigateway_create_custom_domaim")
+    def validate_aws_apigateway_create_custom_domaim(cls, v) -> bool:
+        """Validate aws_apigateway_create_custom_domaim"""
+        if v in [None, ""]:
+            return SettingsDefaults.AWS_APIGATEWAY_CREATE_CUSTOM_DOMAIN
         return v
 
     @field_validator("aws_dynamodb_table_id")