+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-## [0.2.4](https://github.com/FullStackWithLawrence/aws-rekognition/compare/v0.2.3...v0.2.4) (2023-12-20)
+## [0.2.5](https://github.com/FullStackWithLawrence/aws-rekognition/compare/v0.2.4...v0.2.5) (2023-12-20)
 
 
 ### Bug Fixes
 
-* add aws key pair with validation and business rules, and unit tests ([e4d0710](https://github.com/FullStackWithLawrence/aws-rekognition/commit/e4d0710a88c3e04945e3227dc791cb0523e259ae))
+* add a Terraform aws_deployed bool environment variable, to track whether we're running in prod ([d3aa46f](https://github.com/FullStackWithLawrence/aws-rekognition/commit/d3aa46fc3a78bf1181d58844ecc667f6a33fd351))
+
+## [0.2.4](https://github.com/FullStackWithLawrence/aws-rekognition/compare/v0.2.3...v0.2.4) (2023-12-20)
+
+### Bug Fixes
+
+- add aws key pair with validation and business rules, and unit tests ([e4d0710](https://github.com/FullStackWithLawrence/aws-rekognition/commit/e4d0710a88c3e04945e3227dc791cb0523e259ae))
 
 ## [0.2.3](https://github.com/FullStackWithLawrence/aws-rekognition/compare/v0.2.2...v0.2.3) (2023-12-20)